--- _data/projects/active-projects.ts
@@ -1,14 +1,0 @@
-import type { Projects } from "./projects";
-
-export const activeList: Projects[] = [
-  {
-    title: "The Novel Platform",
-    href: "https://github.com/skywalkerSam/the-novel-platform",
-    description: "A Place for Everything Poetic.",
-  },
-  {
-    title: "Animations Demo",
-    href: "https://github.com/skywalkerSam/animations-demo",
-    description: "introduction to Animations and 3D Models.",
-  },
-];
