--- _data/projects/ex-projects.ts
@@ -1,64 +1,0 @@
-import type { Projects } from "./projects";
-
-export const exProjectsList: Projects[] = [
-  {
-    title: "WEB DEV 101",
-    href: "https://github.com/skywalkerSam/web-dev-101",
-    description: "introduction to Web Development.",
-  },
-  {
-    title: "JS 101",
-    href: "https://github.com/skywalkerSam/JS-101",
-    description: "JavaScript 101",
-  },
-  {
-    title: "CS 101",
-    href: "https://github.com/skywalkerSam/CS-101",
-    description: "Computer Science 101",
-  },
-  {
-    title: "DB 101",
-    href: "https://github.com/skywalkerSam/DB-101",
-    description: "Database 101",
-  },
-  {
-    title: "Python 101",
-    href: "https://github.com/skywalkerSam/Python-101",
-    description: "introduction to Python 3",
-  },
-  {
-    title: "Linux 101",
-    href: "https://github.com/skywalkerSam/Linux-101",
-    description: "introduction to Linux.",
-  },
-  {
-    title: "The Hello World Project",
-    href: "https://github.com/skywalkerSam/the-hello-world-project",
-    description: 'A "Hello World" for every programming language in existence.',
-  },
-  {
-    title: "Sidewalks",
-    href: "https://github.com/skywalkerSam/Sidewalks",
-    description: "Side Quests O' Mine...",
-  },
-  {
-    title: "GSAP 101",
-    href: "https://github.com/skywalkerSam/gsap-demo",
-    description: "GSAP Demo.",
-  },
-  {
-    title: "Stardate",
-    href: "https://github.com/skywalkerSam/stardate",
-    description: "A better way of dating things.",
-  },
-  {
-    title: "Starboy Blogs",
-    href: "https://github.com/skywalkerSam/Starboy-Blogs",
-    description: "A Starboy Blog.",
-  },
-  {
-    title: "Starboy Projects",
-    href: "https://github.com/skywalkerSam/Starboy-Projects",
-    description: "Recursion 101",
-  },
-];
