--- _data/projects/projects.ts
@@ -1,19 +1,0 @@
-export type Projects = {
-  title: string;
-  href: string;
-  description: string;
-};
-
-export const projectsList: Projects[] = [
-  {
-    title: "skywalkerSam.dev",
-    href: "https://github.com/skywalkerSam/skywalkerSam.dev",
-    description: "My Personal Corner of The Web.",
-  },
-  {
-    title: "image Gallery",
-    href: "https://github.com/skywalkerSam/image-gallery",
-    description:
-      "A Personalized Gallery Application Demo.",
-  },
-];
