+++ src/app/globals.css
@@ -139,3 +139,7 @@ body {
   }
 }
 
+.relaxed {
+  text-align: justify;
+  line-height: 1.7;
+}
