+++ src/app/projects-index.ts
@@ -1,0 +1,204 @@
+export type Projects = {
+  title: string;
+  href: string;
+  description: string;
+};
+
+export const activeProjects: Projects[] = [
+  {
+    title: "The Novel Platform",
+    href: "https://github.com/skywalkerSam/the-novel-platform",
+    description: "A Place for Everything Poetic.",
+  },
+];
+
+export const personalProjects: Projects[] = [
+  {
+    title: "skywalkerSam.dev",
+    href: "https://github.com/skywalkerSam/skywalkerSam.dev",
+    description: "My Personal Corner of The Web",
+  },
+  {
+    title: "Starboy Blogs",
+    href: "https://github.com/skywalkerSam/Starboy-Blogs",
+    description: "A Starboy's Blog",
+  },
+  {
+    title: "Starboy Projects",
+    href: "https://github.com/skywalkerSam/Starboy-Projects",
+    description: "What is Recursion?",
+  },
+  {
+    title: "Stardate",
+    href: "https://github.com/skywalkerSam/stardate",
+    description: "A better way of dating things.",
+  },
+];
+
+export const oneOones: Projects[] = [
+  {
+    title: "Libraries & Frameworks 101",
+    href: "https://github.com/skywalkerSam/Libraries-and-Frameworks-101",
+    description: "introduction to Libraries & Frameworks w/ freeCodeCamp.org",
+  },
+  {
+    title: "JS 101",
+    href: "https://github.com/skywalkerSam/JS-101",
+    description: "introduction to JavaScript w/ freeCodeCamp.org",
+  },
+  {
+    title: "Responsive Web Design",
+    href: "https://github.com/skywalkerSam/Responsive-Web-Design",
+    description: "Responsive Web Design w/ freeCodeCamp.org",
+  },
+  {
+    title: "GSAP 101",
+    href: "https://github.com/skywalkerSam/gsap-demo",
+    description: "GSAP Demo",
+  },
+  {
+    title: "CS 101",
+    href: "https://github.com/skywalkerSam/CS-101",
+    description: "Computer Science 101",
+  },
+  {
+    title: "DB 101",
+    href: "https://github.com/skywalkerSam/DB-101",
+    description: "Database 101",
+  },
+  {
+    title: "WEB DEV 101",
+    href: "https://github.com/skywalkerSam/web-dev-101",
+    description: "introduction to Web Development",
+  },
+  {
+    title: "Game Development",
+    href: "https://github.com/skywalkerSam/Game-Dev-101",
+    description: "Game Development 101",
+  },
+  {
+    title: "IT 101",
+    href: "https://github.com/skywalkerSam/IT-101",
+    description: "Getting Started with information Technologies",
+  },
+  {
+    title: "ML 101",
+    href: "https://github.com/skywalkerSam/ML-101",
+    description: "Machine Learning 101",
+  },
+  {
+    title: "Python 101",
+    href: "https://github.com/skywalkerSam/Python-101",
+    description: "introduction to Python 3",
+  },
+  {
+    title: "Linux 101",
+    href: "https://github.com/skywalkerSam/Linux-101",
+    description: "introduction to Linux",
+  },
+];
+
+export const archives: Projects[] = [
+  {
+    title: "Animations Demo",
+    href: "https://github.com/skywalkerSam/animations-demo",
+    description: "introduction to Animations and 3D Models",
+  },
+  {
+    title: "image Gallery",
+    href: "https://github.com/skywalkerSam/image-gallery",
+    description: "A Personalized Gallery Application Demo",
+  },
+  {
+    title: "MySpace Demo",
+    href: "https://github.com/skywalkerSam/mySpace-demo",
+    description: "My Space Demo w/ NEXT.js + Vercel",
+  },
+  {
+    title: "Dashboard Demo",
+    href: "https://github.com/skywalkerSam/dashboard-demo",
+    description: "Dashboard Demo w/ NEXT.js + Vercel",
+  },
+  {
+    title: "House Of Cards",
+    href: "https://github.com/skywalkerSam/houseOfCards",
+    description:
+      "A simple game where you'll waste your time matching two cards at a time.)",
+  },
+  {
+    title: "Tic Tac Toe",
+    href: "https://github.com/skywalkerSam/tictactoe",
+    description: "A Simple Tic Tac Toe",
+  },
+  {
+    title: "Animal Farm",
+    href: "https://github.com/skywalkerSam/animalSearch",
+    description: "POC for Searching Things",
+  },
+  {
+    title: "Face Detection",
+    href: "https://github.com/skywalkerSam/faceDetection",
+    description: "Basic Face Detection w/ Clarifai",
+  },
+  {
+    title: "Dall-E Wrapper",
+    href: "https://github.com/skywalkerSam/dalleWrapper",
+    description:
+      "Mimicking Dall-E-like responses with a randomized set of images.",
+  },
+  {
+    title: "Workspace",
+    href: "https://github.com/skywalkerSam/workspace",
+    description: "A Simple Workspace made w/ React",
+  },
+  {
+    title: "Timer",
+    href: "https://github.com/skywalkerSam/timer",
+    description: "A Simple Timer made w/ React",
+  },
+  {
+    title: "TODO",
+    href: "https://github.com/skywalkerSam/todo",
+    description: "A Simple TODO made w/ React",
+  },
+  {
+    title: "Clock",
+    href: "https://github.com/skywalkerSam/clock",
+    description: "A Simple Clock made w/ React",
+  },
+  {
+    title: "Robofriends",
+    href: "https://github.com/skywalkerSam/robofriends",
+    description: "A Collection of Robot Cards made w/ React",
+  },
+  {
+    title: "Hello React",
+    href: "https://github.com/skywalkerSam/hello-react",
+    description: "The Hello World of React",
+  },
+  {
+    title: "The Hello World Project",
+    href: "https://github.com/skywalkerSam/the-hello-world-project",
+    description: 'A "Hello World" for every programming language in existence.',
+  },
+  {
+    title: "Guessing Game",
+    href: "https://github.com/skywalkerSam/Guessing-Game",
+    description: "A Terminal-based Guessing Game written in C++",
+  },
+  {
+    title: "Audio Generator",
+    href: "https://github.com/skywalkerSam/Audio-Generator",
+    description: "Text-to-Audio w/ Python 3",
+  },
+  {
+    title: "Project: Cortana",
+    href: "https://github.com/skywalkerSam/Project-Cortana",
+    description: "Project Cortana: inspired by Halo Infinite",
+  },
+  {
+    title: "Sidewalks",
+    href: "https://github.com/skywalkerSam/Sidewalks",
+    description: "Side Quests O' Mine...",
+  },
+];
